+++ test/passing/tests/module.ml
@@ -39,8 +39,7 @@ end
 module O : sig
   type t
 end
-with type t := t
- and type s := s = struct
+with type t := t and type s := s = struct
   let () = ()
 end
 
@@ -56,10 +55,10 @@ let x : (module S) = (module struct end)
 
 let x = (module struct end : S)
 
-module rec A : (sig
+module rec A : sig
   type t
 end
-with type t = int) = struct
+with type t = int = struct
   type t = int
 end