+++ test/passing/tests/module_type.ml
@@ -24,43 +24,31 @@ module type BAR = sig
   and B : FOO
 end
 
-module type M =
-  module type of M
-    with module A := A
-    (*test*)
-     and module A = A
-    (*test*)
-     and module A = A
-    with module A = A
-    (*test*)
-    with module A = A
+module type M = module type of M
+  with module A := A (*test*) and module A = A (*test*) and module A = A
+  with module A = A (*test*) with module A = A
 
-module U :
-  S
-    with type ttttttttt = int
-     and type uuuuuuuu = int
-     and type vvvvvvvvvvv = int = struct end
+module U : S
+  with type ttttttttt = int
+   and type uuuuuuuu = int
+   and type vvvvvvvvvvv = int = struct end
 
-module U :
-  S
-    with type ttttttttt = int
-     and type uuuuuuu = int
-    with type vvvvvvvvv = int = struct end
+module U : S with type ttttttttt = int and type uuuuuuu = int
+  with type vvvvvvvvv = int = struct end
 
-module U :
-  S
-    with type Command.t =
-      [ `Halt
-      | `Unknown
-      | `Error of string
-      | `Config of (string * string) list
-      | `Format of string ]
-     and type Command.t =
-      [ `Halt
-      | `Unknown
-      | `Error of string
-      | `Config of (string * string) list
-      | `Format of string ] = struct end
+module U : S
+  with type Command.t =
+    [ `Halt
+    | `Unknown
+    | `Error of string
+    | `Config of (string * string) list
+    | `Format of string ]
+   and type Command.t =
+    [ `Halt
+    | `Unknown
+    | `Error of string
+    | `Config of (string * string) list
+    | `Format of string ] = struct end
 
 module U = (val S : S with type t = int and type u = int)
 
@@ -73,15 +61,13 @@ module type S = sig
 end
 
 module type S' = functor
-  (A : A)
-  (B : sig
-     type t
-   end)
-  (Cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
-   : sig
-     type t
-   end)
-  -> S with type t = B.t
+    (A : A)
+    (B : sig
+       type t
+     end)
+    (Cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc : sig
+       type t
+     end) -> S with type t = B.t
 
 module M : sig
   include (* foo *) module type of K
@@ -90,7 +76,8 @@ module M : sig
       Fooooooooooooooooooooooooooo (Foooooooooo.Foo) (Fooooooooooooo)
         (Fooooooooooooo)
 
-  include (* fooooooooo *) module type of
+  include (* fooooooooo *)
+    module type of
       Fooooooooooooooooooooooooooo (Foooooooooo.Foo) (Fooooooooooooo)
         (Fooooooooooooo)
 end = struct end